+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Changed
 
+* Bump `atomic-destructor` to `v0.2.0-alpha.2` ([Yuki Kishimoto])
 * nostr: ignore malformed public keys during NIP19 event (`nevent`) parsing ([Yuki Kishimoto])
 * pool: increase default kind 3 event limit to `840000` bytes and `10000` tags ([Yuki Kishimoto])